@@ -1,7 +1,10 @@
 import {Component, OnInit} from '@angular/core';
-import {FormControl, FormGroup} from "@angular/forms";
-import {ChartAppearanceService} from "../../../common/services/share-services/chart-apperance/chart-appearance.service";
-import {tap} from "rxjs";
+import {UntypedFormControl, UntypedFormGroup} from "@angular/forms";
+import {
+  chartAppearance,
+  ChartAppearanceService
+} from "../../../common/services/share-services/chart-apperance/chart-appearance.service";
+import {debounceTime, distinctUntilChanged, first, tap} from "rxjs";
 
 @Component({
   selector: 'app-chart-options',
@@ -9,18 +12,22 @@ import {tap} from "rxjs";
   styleUrls: ['./chart-options.component.scss']
 })
 export class ChartOptionsComponent implements OnInit {
-  chartOptionsForm = new FormGroup({
-    dryAdiabaticFunctionColor: new FormControl(''),
-
+  chartOptionsForm = new UntypedFormGroup({
+    dryAdiabaticFunctionColor: new UntypedFormControl(''),
+    saturatedAdiabaticFunctionColor: new UntypedFormControl('')
   });
-  defaultOptions$ =this.chartAppearance.chartAppearance$.pipe(
+  loadedDefaultOptions$ = this.chartAppearance.chartAppearance$.pipe(
+    first(),
     tap((options) => {
       this.chartOptionsForm.patchValue(options)
-      console.log(options)
     })
   )
 
   constructor(private readonly chartAppearance: ChartAppearanceService) {
+    this.chartOptionsForm.valueChanges.pipe(debounceTime(400), distinctUntilChanged()).subscribe(newValues => {
+      this.chartAppearance.setChartAppearance(newValues as chartAppearance)
+    })
+
   }
 
   ngOnInit(): void {
