@@ -4,6 +4,8 @@ import {BehaviorSubject} from "rxjs";
 export interface chartAppearance {
   // moistLineColor: string,
   dryAdiabaticFunctionColor: string,
+  dryAdiabaticFunctionSize:number
+  saturatedAdiabaticFunctionColor: string,
   // thermoDataColor: string
   // dewDataColor: string,
   // ratioLineColor: string,
@@ -13,13 +15,22 @@ export interface chartAppearance {
   providedIn: 'root'
 })
 export class ChartAppearanceService {
-  private chartAppearance = new BehaviorSubject<chartAppearance>({dryAdiabaticFunctionColor: '#347957'});
+  private chartAppearance = new BehaviorSubject<chartAppearance>(
+    DEFAULT_APPERANCE
+  );
   chartAppearance$ = this.chartAppearance.asObservable();
 
   constructor() {
   }
 
-  setChartAppearance(chartAppearance:chartAppearance) {
+  setChartAppearance(chartAppearance: chartAppearance) {
     this.chartAppearance.next(chartAppearance)
   }
 }
+
+export let DEFAULT_APPERANCE = {
+  dryAdiabaticFunctionColor: '#347957',
+  dryAdiabaticFunctionSize:6,
+  saturatedAdiabaticFunctionColor: '#2f7ed8'
+
+} as chartAppearance
